We can now delete only one stream instead of all of them

+++ delete_streams.py
@@ -2,9 +2,13 @@ import boto3
 import argparse
 
 def create_parser():
-    parser = argparse.ArgumentParser()
+    parser = argparse.ArgumentParser("""
+Delete all streams, or if a stream is specified, only delete that one
+""")
     parser.add_argument("-r", "--regionName", "--region", dest="region", default="us-east-1",
-                        help="The region you'd like to make this stream in. Default is 'us-east-1'", metavar="REGION_NAME",)
+                        help="The region you'd like to make this stream in. Default is 'us-east-1'.", metavar="REGION_NAME",)
+    parser.add_argument("-s", "--stream", dest="stream_name", default=None,
+                        help="The stream you'd like to delete. If no stream is selected, delete all.", metavar="STREAM_NAME",)
     return parser.parse_args()
 
 def main():
@@ -15,6 +19,8 @@ def main():
     
     # Delete all stream
     for stream_name in response["StreamNames"]:
+        if args.stream_name is not None and args.stream_name != stream_name:
+            continue
         print("Deleting stream {}".format(stream_name))
         kinesis_client.delete_stream(StreamName=stream_name)
